Large refactor to permit the creation of all worksheets in-memory.
This change creates all worksheet files in-memory and only writes them
to disk once the XLSX file is closed.

A new option for creating a version that is in-memory only is available
with `Worksheet::create_in_memory()` which returns the buffer holding
the completed XLSX file contents when closed.

Cleaned up numerous `cargo clippy` lints.